requirements.txt
streamlit>=1.31
pandas
google-cloud-bigquery
llama-index-core